TeX booklet: v1.1, supports all pagecounts
The booklet.tex script now includes a patch by Jakob Lund, sent by mail
on 2011/04/04, that lets the script work for any pagesize, not just
multiples of four. The script is now also pseudo-versioned, the latest
one being version 1.1. The index page is updated to feature download
links for this latest version and version 1.0, which is how we've
back-versioned the previous TeX script.